Rename to better reflect intent
