+++ source/Messaging.Application/Transactions/MoveIn/ICustomerMasterDataClient.cs
@@ -19,12 +19,12 @@ namespace Messaging.Application.Transactions.MoveIn;
 /// <summary>
 /// Interface for fetching customer master data
 /// </summary>
-public interface IRequestCustomerMasterData
+public interface ICustomerMasterDataClient
 {
     /// <summary>
     /// Request master data for a customer
     /// </summary>
     /// <param name="fetchCustomerMasterData"></param>
     /// <returns><see cref="Task"/></returns>
-    Task RequestMasterDataForAsync(FetchCustomerMasterData fetchCustomerMasterData);
+    Task RequestAsync(FetchCustomerMasterData fetchCustomerMasterData);
 }
